+++ src/org/ojalgo/matrix/store/GenericDenseStore.java
@@ -57,7 +57,7 @@ import org.ojalgo.type.context.NumberContext;
  */
 public final class GenericDenseStore<N extends Number & Scalar<N>> extends ScalarArray<N> implements PhysicalStore<N>, DecompositionStore<N> {
 
-    public static interface GenericMultiplyBoth<N extends Number & Scalar<N>> extends ElementsConsumer.FillByMultiplying<N> {
+    public static interface GenericMultiplyBoth<N extends Number & Scalar<N>> extends TransformableRegion.FillByMultiplying<N> {
 
     }
 
@@ -980,24 +980,24 @@ public final class GenericDenseStore<N extends Number & Scalar<N>> extends Scala
         return myFactory;
     }
 
-    public final ElementsConsumer<N> regionByColumns(final int... columns) {
-        return new ElementsConsumer.ColumnsRegion<>(this, multiplyBoth, columns);
+    public final TransformableRegion<N> regionByColumns(final int... columns) {
+        return new TransformableRegion.ColumnsRegion<>(this, multiplyBoth, columns);
     }
 
-    public final ElementsConsumer<N> regionByLimits(final int rowLimit, final int columnLimit) {
-        return new ElementsConsumer.LimitRegion<>(this, multiplyBoth, rowLimit, columnLimit);
+    public final TransformableRegion<N> regionByLimits(final int rowLimit, final int columnLimit) {
+        return new TransformableRegion.LimitRegion<>(this, multiplyBoth, rowLimit, columnLimit);
     }
 
-    public final ElementsConsumer<N> regionByOffsets(final int rowOffset, final int columnOffset) {
-        return new ElementsConsumer.OffsetRegion<>(this, multiplyBoth, rowOffset, columnOffset);
+    public final TransformableRegion<N> regionByOffsets(final int rowOffset, final int columnOffset) {
+        return new TransformableRegion.OffsetRegion<>(this, multiplyBoth, rowOffset, columnOffset);
     }
 
-    public final ElementsConsumer<N> regionByRows(final int... rows) {
-        return new ElementsConsumer.RowsRegion<>(this, multiplyBoth, rows);
+    public final TransformableRegion<N> regionByRows(final int... rows) {
+        return new TransformableRegion.RowsRegion<>(this, multiplyBoth, rows);
     }
 
-    public final ElementsConsumer<N> regionByTransposing() {
-        return new ElementsConsumer.TransposedRegion<>(this, multiplyBoth);
+    public final TransformableRegion<N> regionByTransposing() {
+        return new TransformableRegion.TransposedRegion<>(this, multiplyBoth);
     }
 
     public void rotateRight(final int low, final int high, final double cos, final double sin) {
@@ -1083,7 +1083,7 @@ public final class GenericDenseStore<N extends Number & Scalar<N>> extends Scala
         }
     }
 
-    public void supplyTo(final ElementsConsumer<N> receiver) {
+    public void supplyTo(final TransformableRegion<N> receiver) {
         receiver.fillMatching(this);
     }