+++ src/org/ojalgo/matrix/store/IdentityStore.java
@@ -83,9 +83,9 @@ final class IdentityStore<N extends Number> extends FactoryStore<N> {
     }
 
     @SuppressWarnings("unchecked")
-    public void multiply(final Access1D<N> right, final ElementsConsumer<N> target) {
+    public void multiply(final Access1D<N> right, final TransformableRegion<N> target) {
         if (right instanceof Access2D.Collectable) {
-            ((Access2D.Collectable<N, ElementsConsumer<N>>) right).supplyTo(target);
+            ((Access2D.Collectable<N, TransformableRegion<N>>) right).supplyTo(target);
         } else {
             super.multiply(right, target);
         }
@@ -125,7 +125,7 @@ final class IdentityStore<N extends Number> extends FactoryStore<N> {
         return super.premultiply(left);
     }
 
-    public void supplyTo(final ElementsConsumer<N> receiver) {
+    public void supplyTo(final TransformableRegion<N> receiver) {
 
         receiver.reset();