+++ src/org/ojalgo/structure/Mutate2D.java
@@ -42,104 +42,104 @@ public interface Mutate2D extends Structure2D, Mutate1D {
      */
     interface Exchangeable extends Structure2D {
 
-        void exchangeColumns(final long colA, final long colB);
+        void exchangeColumns(long colA, long colB);
 
-        void exchangeRows(final long rowA, final long rowB);
+        void exchangeRows(long rowA, long rowB);
 
     }
 
     interface Fillable<N extends Number> extends Structure2D, Mutate1D.Fillable<N> {
 
-        default void fillColumn(final long col, final Access1D<N> values) {
+        default void fillColumn(long col, Access1D<N> values) {
             this.fillColumn(0L, col, values);
         }
 
-        default void fillColumn(final long row, final long col, final Access1D<N> values) {
+        default void fillColumn(long row, long col, Access1D<N> values) {
             this.loopColumn(row, col, (r, c) -> this.fillOne(r, c, values.get(r - row)));
         }
 
-        default void fillColumn(final long row, final long col, final N value) {
+        default void fillColumn(long row, long col, N value) {
             this.loopColumn(row, col, (r, c) -> this.fillOne(r, c, value));
         }
 
-        default void fillColumn(final long row, final long col, final NullaryFunction<N> supplier) {
+        default void fillColumn(long row, long col, NullaryFunction<N> supplier) {
             this.loopColumn(row, col, (r, c) -> this.fillOne(r, c, supplier));
         }
 
-        default void fillColumn(final long col, final N value) {
+        default void fillColumn(long col, N value) {
             this.fillColumn(0L, col, value);
         }
 
-        default void fillColumn(final long col, final NullaryFunction<N> supplier) {
+        default void fillColumn(long col, NullaryFunction<N> supplier) {
             this.fillColumn(0L, col, supplier);
         }
 
-        default void fillDiagonal(final Access1D<N> values) {
+        default void fillDiagonal(Access1D<N> values) {
             this.fillDiagonal(0L, 0L, values);
         }
 
-        default void fillDiagonal(final long row, final long col, final Access1D<N> values) {
+        default void fillDiagonal(long row, long col, Access1D<N> values) {
             this.loopDiagonal(row, col, (r, c) -> this.fillOne(r, c, values.get(r - row)));
         }
 
-        default void fillDiagonal(final long row, final long col, final N value) {
+        default void fillDiagonal(long row, long col, N value) {
             this.loopDiagonal(row, col, (r, c) -> this.fillOne(r, c, value));
         }
 
-        default void fillDiagonal(final long row, final long col, final NullaryFunction<N> supplier) {
+        default void fillDiagonal(long row, long col, NullaryFunction<N> supplier) {
             this.loopDiagonal(row, col, (r, c) -> this.fillOne(r, c, supplier));
         }
 
-        default void fillDiagonal(final N value) {
+        default void fillDiagonal(N value) {
             this.fillDiagonal(0L, 0L, value);
         }
 
-        default void fillDiagonal(final NullaryFunction<N> supplier) {
+        default void fillDiagonal(NullaryFunction<N> supplier) {
             this.fillDiagonal(0L, 0L, supplier);
         }
 
-        default void fillOne(final long index, final Access1D<?> values, final long valueIndex) {
-            final long tmpStructure = this.countRows();
+        default void fillOne(long index, Access1D<?> values, long valueIndex) {
+            long tmpStructure = this.countRows();
             this.fillOne(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), values, valueIndex);
         }
 
-        void fillOne(long row, long col, final Access1D<?> values, final long valueIndex);
+        void fillOne(long row, long col, Access1D<?> values, long valueIndex);
 
         void fillOne(long row, long col, N value);
 
         void fillOne(long row, long col, NullaryFunction<N> supplier);
 
-        default void fillOne(final long index, final N value) {
-            final long tmpStructure = this.countRows();
+        default void fillOne(long index, N value) {
+            long tmpStructure = this.countRows();
             this.fillOne(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), value);
         }
 
-        default void fillOne(final long index, final NullaryFunction<N> supplier) {
-            final long tmpStructure = this.countRows();
+        default void fillOne(long index, NullaryFunction<N> supplier) {
+            long tmpStructure = this.countRows();
             this.fillOne(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), supplier);
         }
 
-        default void fillRow(final long row, final Access1D<N> values) {
+        default void fillRow(long row, Access1D<N> values) {
             this.fillRow(row, 0L, values);
         }
 
-        default void fillRow(final long row, final long col, final Access1D<N> values) {
+        default void fillRow(long row, long col, Access1D<N> values) {
             this.loopRow(row, col, (r, c) -> this.fillOne(r, c, values.get(c - col)));
         }
 
-        default void fillRow(final long row, final long col, final N value) {
+        default void fillRow(long row, long col, N value) {
             this.loopRow(row, col, (r, c) -> this.fillOne(r, c, value));
         }
 
-        default void fillRow(final long row, final long col, final NullaryFunction<N> supplier) {
+        default void fillRow(long row, long col, NullaryFunction<N> supplier) {
             this.loopRow(row, col, (r, c) -> this.fillOne(r, c, supplier));
         }
 
-        default void fillRow(final long row, final N value) {
+        default void fillRow(long row, N value) {
             this.fillRow(row, 0L, value);
         }
 
-        default void fillRow(final long row, final NullaryFunction<N> supplier) {
+        default void fillRow(long row, NullaryFunction<N> supplier) {
             this.fillRow(row, 0L, supplier);
         }
 
@@ -147,13 +147,13 @@ public interface Mutate2D extends Structure2D, Mutate1D {
 
     interface Mixable<N extends Number> extends Structure2D, Mutate1D.Mixable<N> {
 
-        default double mix(final long index, final BinaryFunction<N> mixer, final double addend) {
-            final long structure = this.countRows();
+        default double mix(long index, BinaryFunction<N> mixer, double addend) {
+            long structure = this.countRows();
             return this.mix(Structure2D.row(index, structure), Structure2D.column(index, structure), mixer, addend);
         }
 
-        default N mix(final long index, final BinaryFunction<N> mixer, final N addend) {
-            final long structure = this.countRows();
+        default N mix(long index, BinaryFunction<N> mixer, N addend) {
+            long structure = this.countRows();
             return this.mix(Structure2D.row(index, structure), Structure2D.column(index, structure), mixer, addend);
         }
 
@@ -164,50 +164,50 @@ public interface Mutate2D extends Structure2D, Mutate1D {
 
     interface Modifiable<N extends Number> extends Structure2D, Mutate1D.Modifiable<N> {
 
-        default void modifyColumn(final long row, final long col, final UnaryFunction<N> modifier) {
+        default void modifyColumn(long row, long col, UnaryFunction<N> modifier) {
             this.loopColumn(row, col, (r, c) -> this.modifyOne(r, c, modifier));
         }
 
-        default void modifyColumn(final long col, final UnaryFunction<N> modifier) {
+        default void modifyColumn(long col, UnaryFunction<N> modifier) {
             this.modifyColumn(0L, col, modifier);
         }
 
-        default void modifyDiagonal(final long row, final long col, final UnaryFunction<N> modifier) {
+        default void modifyDiagonal(long row, long col, UnaryFunction<N> modifier) {
             this.loopDiagonal(row, col, (r, c) -> this.modifyOne(r, c, modifier));
         }
 
-        default void modifyDiagonal(final UnaryFunction<N> modifier) {
+        default void modifyDiagonal(UnaryFunction<N> modifier) {
             this.modifyDiagonal(0L, 0L, modifier);
         }
 
-        default void modifyMatchingInColumns(final Access1D<N> left, final BinaryFunction<N> function) {
+        default void modifyMatchingInColumns(Access1D<N> left, BinaryFunction<N> function) {
             left.loopAll(r -> this.modifyRow(r, function.first(left.get(r))));
         }
 
-        default void modifyMatchingInColumns(final BinaryFunction<N> function, final Access1D<N> right) {
+        default void modifyMatchingInColumns(BinaryFunction<N> function, Access1D<N> right) {
             right.loopAll(r -> this.modifyRow(r, function.second(right.get(r))));
         }
 
-        default void modifyMatchingInRows(final Access1D<N> left, final BinaryFunction<N> function) {
+        default void modifyMatchingInRows(Access1D<N> left, BinaryFunction<N> function) {
             left.loopAll(c -> this.modifyColumn(c, function.first(left.get(c))));
         }
 
-        default void modifyMatchingInRows(final BinaryFunction<N> function, final Access1D<N> right) {
+        default void modifyMatchingInRows(BinaryFunction<N> function, Access1D<N> right) {
             right.loopAll(c -> this.modifyColumn(c, function.second(right.get(c))));
         }
 
         void modifyOne(long row, long col, UnaryFunction<N> modifier);
 
-        default void modifyOne(final long index, final UnaryFunction<N> modifier) {
-            final long tmpStructure = this.countRows();
+        default void modifyOne(long index, UnaryFunction<N> modifier) {
+            long tmpStructure = this.countRows();
             this.modifyOne(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), modifier);
         }
 
-        default void modifyRow(final long row, final long col, final UnaryFunction<N> modifier) {
+        default void modifyRow(long row, long col, UnaryFunction<N> modifier) {
             this.loopRow(row, col, (r, c) -> this.modifyOne(r, c, modifier));
         }
 
-        default void modifyRow(final long row, final UnaryFunction<N> modifier) {
+        default void modifyRow(long row, UnaryFunction<N> modifier) {
             this.modifyRow(row, 0L, modifier);
         }
 
@@ -218,13 +218,15 @@ public interface Mutate2D extends Structure2D, Mutate1D {
      *
      * @author apete
      */
-    interface ModifiableReceiver<N extends Number> extends Modifiable<N>, Receiver<N> {
+    interface ModifiableReceiver<N extends Number> extends Modifiable<N>, Receiver<N>, Exchangeable {
+
+        void modifyAny(Transformation2D<N> modifier);
 
     }
 
     interface Receiver<N extends Number> extends Mutate2D, Fillable<N>, Consumer<Access2D<?>> {
 
-        default void accept(final Access2D<?> supplied) {
+        default void accept(Access2D<?> supplied) {
             if (this.isAcceptable(supplied)) {
                 supplied.loopAll((r, c) -> this.set(r, c, supplied.get(r, c)));
             } else {
@@ -232,34 +234,28 @@ public interface Mutate2D extends Structure2D, Mutate1D {
             }
         }
 
-        default boolean isAcceptable(final Structure2D supplier) {
+        default boolean isAcceptable(Structure2D supplier) {
             return (this.countRows() >= supplier.countRows()) && (this.countColumns() >= supplier.countColumns());
         }
 
     }
 
-    interface Transformable<N extends Number> extends ModifiableReceiver<N> {
-
-        void transform(Transformation2D<N> transformation);
-
-    }
-
-    default void add(final long index, final double addend) {
-        final long tmpStructure = this.countRows();
-        this.add(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), addend);
+    default void add(long index, double addend) {
+        long structure = this.countRows();
+        this.add(Structure2D.row(index, structure), Structure2D.column(index, structure), addend);
     }
 
     void add(long row, long col, double addend);
 
     void add(long row, long col, Number addend);
 
-    default void add(final long index, final Number addend) {
-        final long tmpStructure = this.countRows();
+    default void add(long index, Number addend) {
+        long tmpStructure = this.countRows();
         this.add(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), addend);
     }
 
-    default void set(final long index, final double addend) {
-        final long tmpStructure = this.countRows();
+    default void set(long index, double addend) {
+        long tmpStructure = this.countRows();
         this.set(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), addend);
     }
 
@@ -267,8 +263,8 @@ public interface Mutate2D extends Structure2D, Mutate1D {
 
     void set(long row, long col, Number value);
 
-    default void set(final long index, final Number addend) {
-        final long tmpStructure = this.countRows();
+    default void set(long index, Number addend) {
+        long tmpStructure = this.countRows();
         this.set(Structure2D.row(index, tmpStructure), Structure2D.column(index, tmpStructure), addend);
     }