+++ src/org/ojalgo/structure/Stream2D.java
@@ -29,6 +29,8 @@ public interface Stream2D<N extends Number, A extends Access2D<N>, R extends Mut
 
     P operateOnAll(final UnaryFunction<N> operator);
 
+    P operateOnAny(final Transformation2D<N> operator);
+
     P operateOnMatching(final A left, final BinaryFunction<N> operator);
 
     P operateOnMatching(final BinaryFunction<N> operator, final A right);